Refactor fhir_to_phdc endpoint introducing transform_to_phdc_input_data()
