+++ containers/message-parser/app/utils.py
@@ -9,6 +9,10 @@ from typing import Literal
 import fhirpathpy
 import requests
 from app.config import get_settings
+from app.phdc.models import PHDCInputData
+from app.phdc.models import Patient
+from app.phdc.models import Address
+from app.phdc.models import Name
 from frozendict import frozendict
 
 from phdi.cloud.azure import AzureCredentialManager
@@ -280,3 +284,150 @@ def read_file_from_assets(filename: str):
 
 def get_datetime_now():
     return datetime.datetime.now()
+
+
+def extract_and_apply_parsers(parsing_schema, message, response):
+    """
+    Helper function used to pull parsing methods for each field out of the
+    passed-in schema, resolve any reference dependencies, and apply the
+    result to the input FHIR bundle. If reference dependencies are present
+    (e.g. an Observation resource that references an ordering provider
+    Organization), this function will raise an error if those references
+    cannot be resolved (if the ID of the referenced object can't be found,
+    for example).
+
+    :param parsing_schema: A dictionary holding the parsing schema send
+      to the endpoint.
+    :param message: The FHIR bundle to extract values from.
+    :param response: The Response object the endpoint will send back, in
+      case we need to apply error status codes.
+    :return: A dictionary mapping schema keys to parsed values.
+    """
+    parsers = get_parsers(parsing_schema)
+    parsed_values = {}
+
+    # Iterate over each parser and make the appropriate path call
+    for field, parser in parsers.items():
+        if "secondary_parsers" not in parser:
+            value = parser["primary_parser"](message)
+            if len(value) == 0:
+                value = None
+            else:
+                value = ",".join(map(str, value))
+            parsed_values[field] = value
+
+        # Use the secondary field data structure, remembering that some
+        # fhir paths might not be compiled yet
+        else:
+            inital_values = parser["primary_parser"](message)
+            values = []
+            for initial_value in inital_values:
+                value = {}
+                for secondary_field, path_struct in parser["secondary_parsers"].items():
+                    # Base cases for a secondary field:
+                    # Information is contained on this resource, just in a
+                    # nested structure
+                    if "reference_path" not in path_struct:
+                        try:
+                            secondary_parser = path_struct["secondary_fhir_path"]
+                            if len(secondary_parser(initial_value)) == 0:
+                                value[secondary_field] = None
+                            else:
+                                value[secondary_field] = ",".join(
+                                    map(str, secondary_parser(initial_value))
+                                )
+
+                        # By default, fhirpathpy will compile such that *only*
+                        # actual resources can be accessed, rather than data types.
+                        # This is fine for most cases, but sometimes the actual data
+                        # we want is in a list of structs rather than a list of
+                        # resources, such as a list of patient addresses. This
+                        # exception catches that and allows an ordinary property
+                        # search.
+                        except KeyError:
+                            try:
+                                accessors = (
+                                    secondary_parser.parsedPath.get("children")[0]
+                                    .get("text")
+                                    .split(".")[1:]
+                                )
+                                val = initial_value
+                                for acc in accessors:
+                                    if "[" not in acc:
+                                        val = val[acc]
+                                    else:
+                                        sub_acc = acc.split("[")[1].split("]")[0]
+                                        val = val[acc.split("[")[0].strip()][
+                                            int(sub_acc)
+                                        ]
+                                value[secondary_field] = str(val)
+                            except:  # noqa
+                                value[secondary_field] = None
+
+                    # Reference case: information is contained on another
+                    # resource that we have to look up
+                    else:
+                        reference_parser = path_struct["reference_path"]
+                        if len(reference_parser(initial_value)) == 0:
+                            response.status_code = status.HTTP_400_BAD_REQUEST
+                            return {
+                                "message": "Provided `reference_lookup` location does "
+                                "not point to a referencing identifier",
+                                "parsed_values": {},
+                            }
+                        else:
+                            reference_to_find = ",".join(
+                                map(str, reference_parser(initial_value))
+                            )
+
+                            # FHIR references are prefixed with resource type
+                            reference_to_find = reference_to_find.split("/")[-1]
+
+                            # Build the resultant concatenated reference path
+                            reference_path = path_struct["secondary_fhir_path"].replace(
+                                DIBBS_REFERENCE_SIGNIFIER, reference_to_find
+                            )
+                            reference_path = fhirpathpy.compile(reference_path)
+                            referenced_value = reference_path(message)
+                            if len(referenced_value) == 0:
+                                response.status_code = status.HTTP_400_BAD_REQUEST
+                                return {
+                                    "message": "Provided bundle does not contain a "
+                                    "resource matching reference criteria defined "
+                                    "in schema",
+                                    "parsed_values": {},
+                                }
+                            else:
+                                value[secondary_field] = ",".join(
+                                    map(str, referenced_value)
+                                )
+
+                values.append(value)
+            parsed_values[field] = values
+    return parsed_values
+
+
+def transform_to_phdc_input_data(parsed_values: dict) -> PHDCInputData:
+    """
+    Transform the parsed values into a PHDCInputData object.
+
+    :param parsed_values: A dictionary containing the values parsed out of a FHIR
+        bundle.
+    :return: A PHDCInputData object.
+    """
+    # Translate to internal data classes
+    input_data = PHDCInputData()
+    input_data.patient = Patient()
+    for key, value in parsed_values.items():
+        match key:
+            case "patient_address":
+                input_data.patient.address = [Address(**address) for address in value]
+            case "patient_name":
+                input_data.patient.name = [Name(**name) for name in value]
+            case "patient_administrative_gender_code":
+                input_data.patient.administrative_gender_code = value
+            case "patient_birth_time":
+                input_data.patient.birth_time = value
+            case _:
+                pass
+    return input_data
